+++ ydb/core/ymq/actor/error.cpp
@@ -2,16 +2,6 @@
 
 namespace NKikimr::NSQS {
 
-void MakeError(NSQS::TError* error, const TErrorClass& errorClass, const TString& message) {
-    error->SetErrorCode(errorClass.ErrorCode);
-    error->SetStatus(errorClass.HttpStatusCode);
-    if (!message.empty()) {
-        error->SetMessage(message);
-    } else {
-        error->SetMessage(errorClass.DefaultMessage);
-    }
-}
-
 size_t ErrorsCount(const NKikimrClient::TSqsResponse& response, TAPIStatusesCounters* counters) {
 #define RESPONSE_CASE(action)                                           \
     case NKikimrClient::TSqsResponse::Y_CAT(k, action): {               \
