+++ ydb/core/ymq/error/ya.make
@@ -1,0 +1,14 @@
+LIBRARY()
+
+SRCS(
+    error.cpp
+)
+
+PEERDIR(
+    ydb/core/protos
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+)
